refactor classes into own files
